+++ packages/integration-sdk-runtime/test-event-emitters.sh
@@ -1,0 +1,1 @@
+yarn build:dist && node ./testEventEmitters.js