+++ packages/integration-sdk-runtime/test/mocks/mockWriteJsonToPath.js
@@ -1,0 +1,3 @@
+// Prevent tests from creating ./.j1-integration directory
+const fileSystem = require('../../dist/src/fileSystem');
+fileSystem.writeJsonToPath = () => {};