+++ packages/integration-sdk-runtime/test/util/expect.js
@@ -1,0 +1,15 @@
+function expect(result) {
+  return {
+    toBe: (expected) => {
+      if (result != expected) {
+        console.error('Result does not match expected:\n\tResult:');
+        console.error(result);
+        console.error('\tExpected:');
+        console.error(expected);
+        process.exit(1);
+      }
+    },
+  };
+}
+
+module.exports = { expect };